/base/branches/1.2.2 /include/util/archive 修正细微问题
git-svn-id: http://svn01/repositories/r01/base/branches/1.2.2@5900 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ archive/ArchiveBuffer.h
@@ -158,7 +158,7 @@ namespace util
                         return pos_type(-1);
                     }
                 } else {
-                    if (position <= buf_size_) {
+                    if ((size_t)position <= buf_size_) {
                         this->setg(buf_, buf_ + position, buf_ + position);
                         this->setp(buf_ + position, buf_ + buf_size_);
                     } else {
